# specs/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Teams API
  description: API for managing soccer teams, players, trainers, events, and equipment
  version: 1.0.0
  contact:
    name: Teams API Support
    email: support@teams-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.teams.com
    description: Production server
# Authentication is optional for most endpoints, but required for some
security: []
paths:
...
  /api/groups/{groupId}/events:
    get:
      operationId: getAllEvents
      summary: Get all events in a group
      description: Retrieve all events in a specific group
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createEvent
      summary: Create a new event in a group
      description: Create a new event in a specific group
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/groups/{groupId}/events/{id}:
    get:
      operationId: getEventById
      summary: Get event by ID
      description: Retrieve a specific event by ID in a specific group
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateEventById
      summary: Update event by ID
      description: Update a specific event by ID in a specific group
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteEventById
      summary: Delete event by ID
      description: Delete a specific event by ID in a specific group
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Event deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/groups/{groupId}/events/{id}/players/{player_id}/status:
    put:
      operationId: updateInvitationStatus
      summary: Update invitation status
      description: Update the status of a player's invitation. Guardians can update only invitations for their own children and only to accepted or declined.
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: player_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [open, accepted, declined, injured, sick, unavailable]
              required:
                - status
      responses:
        '200':
          description: Invitation status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Event or invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/groups/{groupId}/events/{id}/teams/{teamId}/status:
    put:
      operationId: updateTeamStatus
      summary: Update team selection status
      description: Update the selection status of a single team within an event. Transitioning to selected requires at least minPlayersPerTeam selected players.
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: teamId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [new, selected]
              required:
                - status
      responses:
        '200':
          description: Team status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid status value or team does not meet minimum player requirement for selection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Event or team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LineupPositionAssignment:
      type: object
      properties:
        slotId:
          type: string
          description: Must reference a slot in the Team's selected Formation.
        playerId:
          type: string
          description: Must be one of the Team's selectedPlayers.
      required:
        - slotId
        - playerId
    TeamLineupPeriod:
      type: object
      properties:
        periodNumber:
          type: integer
          minimum: 1
          description: Must not exceed the Event's Playing Mode numberOfPeriods.
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/LineupPositionAssignment'
          description: >-
            Selected players absent here are implicitly benched for this
            period; no explicit bench list is stored.
      required:
        - periodNumber
        - assignments
...
    Invitation:
      type: object
      properties:
        id:
          type: string
          
        playerId:
          type: string
          
        status:
          type: string
          enum: [open, accepted, declined, injured, sick, unavailable]
      required:
        - id
        - playerId
        - status
...
    ShirtAssignment:
      type: object
      properties:
        playerId:
          type: string
          
        shirtNumber:
          type: integer
          minimum: 1
      required:
        - playerId
        - shirtNumber
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        strength:
          type: integer
          minimum: 1
          maximum: 3
          default: 2
        startTime:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
          example: "14:30"
        location:
          type: string
          description: Optional location for the team (overrides event location)
        selectedPlayers:
          type: array
          items:
            type: string
        trainerId:
          type: string
          nullable: true
        responsiblePersonIds:
          type: array
          items:
            type: string
        shirtSetId:
          type: string
          nullable: true
        shirtAssignments:
          type: array
          items:
            $ref: '#/components/schemas/ShirtAssignment'
          nullable: true
        formationId:
          type: string
          nullable: true
          description: >-
            References a Formation in the same group. Only settable while the
            Team's Event has a non-null playingModeId. Fixed for the whole
            match.
        lineup:
          type: array
          items:
            $ref: '#/components/schemas/TeamLineupPeriod'
          nullable: true
          description: One entry per planned period; unlisted periods are unplanned.
      required:
        - id
        - name
        - strength
        - startTime
        - selectedPlayers
    Event:
      type: object
      properties:
        id:
          type: string
        squadId:
          type: string
        name:
          type: string
        date:
          type: string
          format: date
        maxPlayersPerTeam:
          type: integer
          minimum: 1
        minPlayersPerTeam:
          type: integer
          minimum: 1
        location:
          type: string
        playingModeId:
          type: string
          nullable: true
          description: >-
            References a Playing Mode in the same group. Only usable when the
            group has matchPlanningEnabled true. Always optional/resettable.
        matchFormatId:
          type: string
          nullable: true
          description: Squad alias for playingModeId. Both values refer to the same stored field.
        teams:
          type: array
          items:
            $ref: '#/components/schemas/Team'
        invitations:
          type: array
          items:
            $ref: '#/components/schemas/Invitation'
      required:
        - id
        - name
        - date
        - minPlayersPerTeam
        - maxPlayersPerTeam
        - teams
        - invitations
    CreateEventRequest:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        maxPlayersPerTeam:
          type: integer
          minimum: 1
        minPlayersPerTeam:
          type: integer
          minimum: 1
          default: 1
        location:
          type: string
        playingModeId:
          type: string
          nullable: true
          description: >-
            Omit to auto-assign the group's default Playing Mode (if any);
            pass null explicitly to opt out at creation.
        matchFormatId:
          type: string
          nullable: true
          description: Squad alias for playingModeId; takes precedence when both are supplied.
        teams:
          type: array
          items:
            $ref: '#/components/schemas/Team'
      required:
        - name
        - date
        - maxPlayersPerTeam
        - minPlayersPerTeam
    UpdateEventRequest:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        maxPlayersPerTeam:
          type: integer
          minimum: 1
        location:
          type: string
        playingModeId:
          type: string
          nullable: true
          description: >-
            Set to a Playing Mode id from the same group, or null to reset to
            none. Rejected with 400 if the group has matchPlanningEnabled
            false.
        matchFormatId:
          type: string
          nullable: true
          description: Squad alias for playingModeId; takes precedence when both are supplied.
        teams:
          type: array
          items:
            $ref: '#/components/schemas/Team'
        invitations:
          type: array
          items:
            $ref: '#/components/schemas/Invitation'
...